Refined types, refactored some of the messier parts
