+++ package.json
@@ -52,6 +52,7 @@
     "@fortawesome/react-fontawesome": "^0.1.3",
     "@researchgate/react-intersection-observer": "^0.7.3",
     "@types/firebase": "^3.2.1",
+    "@types/jss": "^9.5.7",
     "@types/react": "^16.4.16",
     "@types/react-jss": "^8.6.0",
     "@types/react-redux": "^6.0.7",
@@ -65,6 +66,7 @@
     "express": "^4.16.3",
     "final-form": "^4.10.0",
     "firebase": "^5.5.4",
+    "jss": "^9.8.7",
     "material-ui": "^0.20.2",
     "prettier": "^1.15.3",
     "react-final-form": "^3.6.7",
