+++ src/js/modules/core/components/SubwayIcon.tsx
@@ -2,15 +2,24 @@ import * as React from "react";
 import { Styles } from "react-jss";
 import {ReactNode} from "react";
 import injectSheet from "react-jss/lib/injectSheet";
+import { JssRules } from "../../types";
 
-const styles: Styles = {
+interface SubwayIconStyles<T> extends Styles {
+  SubwayIcon: T
+};
+
+interface Props {
+  color: string;
+  radius: number;
+  children: ReactNode;
+  classes: SubwayIconStyles<string>
+}
+
+const styles: SubwayIconStyles<JssRules> = {
   SubwayIcon: {
-    // @ts-ignore
     backgroundColor: props => props.color,
     fontFamily: "Helvetica, sans-serif",
-    // @ts-ignore
     width: props => (props.radius * 2),
-    // @ts-ignore
     height: props => (props.radius * 2),
     borderRadius: "80px",
     fontSize: "0.62em",
@@ -28,12 +37,7 @@ const styles: Styles = {
     }
   }
 };
-interface Props {
-  color: string;
-  radius: number;
-  children: ReactNode;
-  classes: { [s: string]: string }
-}
+
 const SubwayIcon: React.SFC<Props> = ({ classes, children }) => {
   return (
     <div className={classes.SubwayIcon}>
