+++ src/js/modules/core/components/SubwayLine.tsx
@@ -1,10 +1,13 @@
 import * as React from "react";
 import { Styles } from "react-jss";
 import injectSheet from "react-jss/lib/injectSheet";
-import { Component } from "react";
-import { delay, getRandomInteger } from "../../utils";
 import SubwayStop from "./SubwayStop";
 import { STOPS_COUNT } from "../../constants";
+import { JssRules } from "../../types";
+
+interface SubwayLineStyles<T> extends Styles {
+  SubwayLine: T;
+}
 
 interface SubwayLineProps {
   color: string;
@@ -12,18 +15,17 @@ interface SubwayLineProps {
   stopOffsets: number[];
   lineOffset: number;
   currentStop: number;
-  classes: { [s: string]: string };
+  classes: SubwayLineStyles<string>;
 }
 
-const styles: Styles = {
+const styles: SubwayLineStyles<JssRules> = {
   SubwayLine: {
     height: "15px",
-    // @ts-ignore
-    width: props => `${props.lineStep * (100 / STOPS_COUNT) + props.lineOffset}vw`,
+    width: props =>
+      `${props.lineStep * (100 / STOPS_COUNT) + props.lineOffset}vw`,
     transition: "width 1s",
     margin: "10px 0px 10px 0px",
     borderRadius: "10%",
-    // @ts-ignore
     backgroundColor: props => props.color
   }
 };
@@ -48,76 +50,4 @@ const SubwayLine: React.SFC<SubwayLineProps> = ({
   return <div className={classes.SubwayLine}>{subwayStops}</div>;
 };
 
-const StyledSubwayLine = injectSheet(styles)(SubwayLine);
-
-interface State {
-  lineStep: number;
-  currentStop: number;
-  stopOffsets: number[];
-  lineOffset: number;
-}
-
-interface Props {
-  delay: number;
-  color: string;
-}
-
-class SubwayLineWrapper extends Component<Props, State> {
-  isMounted: boolean;
-
-  constructor(props: Props) {
-    super(props);
-    let stopOffsets = [];
-    for (let i = 0; i < STOPS_COUNT; i++) {
-      stopOffsets.push(getRandomInteger(5) + 3);
-    }
-    this.isMounted = true;
-    this.state = {
-      lineStep: 0,
-      currentStop: 0,
-      lineOffset: getRandomInteger(5) + 3,
-      stopOffsets
-    };
-  }
-
-  componentDidMount() {
-    delay(this.props.delay).then(() => this.animate());
-  }
-
-  componentWillUnmount() {
-    this.isMounted = false;
-  }
-
-  async animate() {
-    while (this.isMounted) {
-      while (this.state.lineStep < STOPS_COUNT) {
-        await delay(500 + getRandomInteger(1000));
-        this.setState({ lineStep: this.state.lineStep + 1 });
-        await delay(750);
-        this.setState({ currentStop: this.state.currentStop + 1 });
-      }
-      await delay(1000);
-      while (this.state.lineStep > 0) {
-        await delay(750);
-        this.setState({ currentStop: this.state.currentStop - 1 });
-        await delay(500 + getRandomInteger(1000));
-        this.setState({ lineStep: this.state.lineStep - 1 });
-      }
-    }
-  }
-
-  render() {
-    const { color } = this.props;
-    const { lineStep, stopOffsets, lineOffset, currentStop } = this.state;
-    return (
-      <StyledSubwayLine
-        lineOffset={lineOffset}
-        color={color}
-        stopOffsets={stopOffsets}
-        lineStep={lineStep}
-        currentStop={currentStop}
-      />
-    );
-  }
-}
-export default SubwayLineWrapper;
+export default injectSheet(styles)(SubwayLine);
